+++ app/src/main/kotlin/com/pitchedapps/frost/injectors/CssAssets.kt
@@ -25,21 +25,19 @@ enum class CssAssets(val folder: String = "themes") : InjectorContract {
             try {
                 var content = webView.context.assets.open("css/$folder/$file").bufferedReader().use { it.readText() }
                 if (this == CUSTOM) {
-                    var bbt = Prefs.bgColor
                     val bt: String
-                    if (Color.alpha(bbt) == 255) {
-                        bbt = bbt.adjustAlpha(0.2f).colorToForeground(0.35f)
+                    if (Color.alpha(Prefs.bgColor) == 255) {
                         bt = Prefs.bgColor.toRgbaString()
                     } else {
-                        bbt = bbt.adjustAlpha(0.05f).colorToForeground(0.5f)
                         bt = "transparent"
                     }
                     content = content
                             .replace("\$T\$", Prefs.textColor.toRgbaString())
                             .replace("\$TT\$", Prefs.textColor.colorToBackground(0.05f).toRgbaString())
+                            .replace("\$A\$", Prefs.accentColor.toRgbaString())
                             .replace("\$B\$", Prefs.bgColor.toRgbaString())
                             .replace("\$BT\$", bt)
-                            .replace("\$BBT\$", bbt.toRgbaString())
+                            .replace("\$BBT\$", Prefs.bgColor.withAlpha(51).colorToForeground(0.35f).toRgbaString())
                             .replace("\$O\$", Prefs.bgColor.withAlpha(255).toRgbaString())
                             .replace("\$OO\$", Prefs.bgColor.colorToForeground(0.35f).withAlpha(255).toRgbaString())
                             .replace("\$D\$", Prefs.textColor.adjustAlpha(0.3f).toRgbaString())