+++ app/src/main/kotlin/com/pitchedapps/frost/web/FrostWebViewClients.kt
@@ -11,7 +11,6 @@ import com.pitchedapps.frost.activities.LoginActivity
 import com.pitchedapps.frost.activities.MainActivity
 import com.pitchedapps.frost.activities.SelectorActivity
 import com.pitchedapps.frost.activities.WebOverlayActivity
-import com.pitchedapps.frost.facebook.FACEBOOK_COM
 import com.pitchedapps.frost.facebook.FB_URL_BASE
 import com.pitchedapps.frost.facebook.FbCookie
 import com.pitchedapps.frost.facebook.FbItem
@@ -34,7 +33,7 @@ import org.jetbrains.anko.withAlpha
 open class BaseWebViewClient : WebViewClient() {
 
     override fun shouldInterceptRequest(view: WebView, request: WebResourceRequest): WebResourceResponse?
-            = shouldFrostInterceptRequest(view, request)
+            = view.shouldFrostInterceptRequest(request)
 
 }
 
@@ -51,7 +50,7 @@ open class FrostWebViewClient(val webCore: FrostWebViewCore) : BaseWebViewClient
         if (url == null) return
         L.i("FWV Loading", url)
         refreshObservable.onNext(true)
-        if (!url.contains(FACEBOOK_COM)) return
+        if (!url.isFacebookUrl) return
         if (url.contains("logout.php")) FbCookie.logout(Prefs.userId, { launchLogin(view.context) })
         else if (url.contains("login.php")) FbCookie.reset({ launchLogin(view.context) })
     }
@@ -71,18 +70,19 @@ open class FrostWebViewClient(val webCore: FrostWebViewCore) : BaseWebViewClient
     override fun onPageCommitVisible(view: WebView, url: String?) {
         super.onPageCommitVisible(view, url)
         injectBackgroundColor()
-        view.jsInject(
-                CssAssets.ROUND_ICONS.maybe(Prefs.showRoundedIcons),
-                CssHider.HEADER,
-                CssHider.PEOPLE_YOU_MAY_KNOW.maybe(!Prefs.showSuggestedFriends && IS_FROST_PRO),
-                Prefs.themeInjector,
-                CssHider.NON_RECENT.maybe(webCore.url?.contains("?sk=h_chr") ?: false))
+        if (url.isFacebookUrl)
+            view.jsInject(
+                    CssAssets.ROUND_ICONS.maybe(Prefs.showRoundedIcons),
+                    CssHider.HEADER,
+                    CssHider.PEOPLE_YOU_MAY_KNOW.maybe(!Prefs.showSuggestedFriends && IS_FROST_PRO),
+                    Prefs.themeInjector,
+                    CssHider.NON_RECENT.maybe(webCore.url?.contains("?sk=h_chr") ?: false))
     }
 
     override fun onPageFinished(view: WebView, url: String?) {
         url ?: return
         L.i("Page finished", url)
-        if (!url.contains(FACEBOOK_COM)) {
+        if (!url.isFacebookUrl) {
             refreshObservable.onNext(false)
             return
         }
@@ -124,9 +124,7 @@ open class FrostWebViewClient(val webCore: FrostWebViewCore) : BaseWebViewClient
      */
     private fun launchRequest(request: WebResourceRequest): Boolean {
         L.d("Launching Url", request.url?.toString() ?: "null")
-        if (webCore.context is WebOverlayActivity) return false
-        webCore.context.launchWebOverlay(request.url.toString())
-        return true
+        return webCore.context !is WebOverlayActivity && webCore.context.requestWebOverlay(request.url.toString())
     }
 
     private fun launchImage(url: String, text: String? = null): Boolean {